# libs/cli/client_api/src/models/callback_model.rs
/*
 * ProxyAuthK8S
 *
 * Reverse Proxy K8s Api.
 *
 * The version of the OpenAPI document: 0.1.0
 * Contact: maxleriche.60@gmail.com
 * Generated by: https://openapi-generator.tech
 */

use crate::models;
use serde::{Deserialize, Serialize};

/// CallbackModel : Model for the callback response after successful authentication with the cluster.  This model contains the access token, refresh token, cluster URL, subject, and ID token returned by the cluster after successful authentication.
#[derive(Clone, Default, Debug, PartialEq, Serialize, Deserialize)]
pub struct CallbackModel {
    #[serde(rename = "access_token")]
    pub access_token: String,
    #[serde(rename = "cluster_url")]
    pub cluster_url: String,
    #[serde(rename = "id_token")]
    pub id_token: String,
    #[serde(rename = "refresh_token")]
    pub refresh_token: String,
    #[serde(rename = "subject")]
    pub subject: String,
}

impl CallbackModel {
    /// Model for the callback response after successful authentication with the cluster.  This model contains the access token, refresh token, cluster URL, subject, and ID token returned by the cluster after successful authentication.
    pub fn new(
        access_token: String,
        cluster_url: String,
        id_token: String,
        refresh_token: String,
        subject: String,
    ) -> CallbackModel {
        CallbackModel {
            access_token,
            cluster_url,
            id_token,
            refresh_token,
            subject,
        }
    }
}