# libs/front/front-api/src/lib/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
    baseURL: `${string}://${string}` | (string & {});
};

/**
 * Model representing a cluster visible to the user.
 */
export type VisibleCluster = {
    enabled: boolean;
    is_reachable?: boolean | null;
    name: string;
    namespace: string;
    sso_enabled: boolean;
};

/**
 * Body of the response for the get_all_visible_cluster endpoint.
 *
 * Contains a list of clusters visible to the user.
 * Will be empty if the user has no clusters visible to them.
 */
export type GetAllVisibleClusterBody = {
    clusters: Array<VisibleCluster>;
};

/**
 * Model for the callback response after successful authentication with the cluster.
 *
 * This model contains the access token, refresh token, cluster URL, subject, and ID token returned by the cluster after successful authentication.
 */
export type CallbackModel = {
    access_token: string;
    cluster_url: string;
    id_token: string;
    refresh_token: string;
    subject: string;
};

export type GetAllVisibleClusterData = {
    body?: never;
    path?: never;
    query?: never;
    url: '/api/v1/clusters';
};

export type GetAllVisibleClusterErrors = {
    /**
     * User is not authenticated.
     */
    401: unknown;
    /**
     * Internal server error.
     */
    500: unknown;
};

export type GetAllVisibleClusterResponses = {
    /**
     * Get all visible clusters.
     */
    200: GetAllVisibleClusterBody;
};

export type GetAllVisibleClusterResponse = GetAllVisibleClusterResponses[keyof GetAllVisibleClusterResponses];

export type CallbackLoginData = {
    body?: never;
    headers: {
        /**
         * If it's from the frontend, this header will be set.
         */
        'x-front-callback': string | null;
        /**
         * If it's from kubectl plugin, this header will be set.
         */
        'x-kubectl-callback': string | null;
    };
    path: {
        /**
         * Namespace containing the cluster.
         */
        ns: string;
        /**
         * Cluster name that should exist in the namespace.
         */
        cluster: string;
    };
    query: {
        /**
         * Authorization code from the OIDC provider.
         */
        code: string;
        /**
         * State parameter to prevent CSRF.
         */
        state: string;
    };
    url: '/clusters/{ns}/{cluster}/auth/callback';
};

export type CallbackLoginErrors = {
    /**
     * Cluster not found or disabled.
     */
    404: unknown;
    /**
     * Internal server error.
     */
    500: unknown;
};

export type CallbackLoginResponses = {
    /**
     * Response from remote cluster.
     */
    200: CallbackModel;
};

export type CallbackLoginResponse = CallbackLoginResponses[keyof CallbackLoginResponses];

export type ClusterLoginData = {
    body?: never;
    headers: {
        /**
         * If it's from the frontend, this header will be set.
         */
        'x-front-callback': string | null;
        /**
         * If it's from kubectl plugin, this header will be set.
         */
        'x-kubectl-callback': string | null;
    };
    path: {
        /**
         * Namespace containing the cluster.
         */
        ns: string;
        /**
         * Cluster name that should exist in the namespace.
         */
        cluster: string;
    };
    query?: never;
    url: '/clusters/{ns}/{cluster}/auth/login';
};

export type ClusterLoginErrors = {
    /**
     * Cluster not found or disabled.
     */
    404: unknown;
    /**
     * Internal server error.
     */
    500: unknown;
};

export type ClusterLoginResponses = {
    /**
     * Response from remote cluster.
     */
    200: string;
};

export type ClusterLoginResponse = ClusterLoginResponses[keyof ClusterLoginResponses];