# swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ProxyAuthK8S",
    "description": "Reverse Proxy K8s Api.",
    "contact": {
      "name": "Batleforc",
      "url": "https://maxleriche.net",
      "email": "maxleriche.60@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current usable server."
    }
  ],
  "paths": {
...
    "/clusters/{ns}/{cluster}/auth/callback": {
      "get": {
        "tags": [
          "auth_clusters"
        ],
        "summary": "Callback from the cluster's OIDC provider",
        "description": "If the cluster is not found or disabled, return 404.",
        "operationId": "callback_login",
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "description": "Namespace containing the cluster.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cluster",
            "in": "path",
            "description": "Cluster name that should exist in the namespace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-front-callback",
            "in": "header",
            "description": "If it's from the frontend, this header will be set.",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "x-kubectl-callback",
            "in": "header",
            "description": "If it's from kubectl plugin, this header will be set.",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "code",
            "in": "query",
            "description": "Authorization code from the OIDC provider.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "state",
            "in": "query",
            "description": "State parameter to prevent CSRF.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response from remote cluster.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CallbackModel"
                }
              }
            }
          },
          "404": {
            "description": "Cluster not found or disabled."
          },
          "500": {
            "description": "Internal server error."
          }
        }
      }
    },
    "/clusters/{ns}/{cluster}/auth/login": {
      "get": {
        "tags": [
          "auth_clusters"
        ],
        "summary": "Redirect to the cluster's login page",
        "description": "If the cluster is not found or disabled, return 404.",
        "operationId": "cluster_login",
        "parameters": [
          {
            "name": "ns",
            "in": "path",
            "description": "Namespace containing the cluster.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cluster",
            "in": "path",
            "description": "Cluster name that should exist in the namespace.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "x-front-callback",
            "in": "header",
            "description": "If it's from the frontend, this header will be set.",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "x-kubectl-callback",
            "in": "header",
            "description": "If it's from kubectl plugin, this header will be set.",
            "required": true,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Response from remote cluster.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Cluster not found or disabled."
          },
          "500": {
            "description": "Internal server error."
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "CallbackModel": {
        "type": "object",
        "description": "Model for the callback response after successful authentication with the cluster.\n\nThis model contains the access token, refresh token, cluster URL, subject, and ID token returned by the cluster after successful authentication.",
        "required": [
          "access_token",
          "refresh_token",
          "cluster_url",
          "subject",
          "id_token"
        ],
        "properties": {
          "access_token": {
            "type": "string"
          },
          "cluster_url": {
            "type": "string"
          },
          "id_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}